requirements.txt
numpy>=1.18.5
torchvision==0.9.1
torch==1.8.1
runstats>=1.8.0
pytorch_lightning==1.3.3
h5py>=2.10.0
PyYAML>=5.3.1
pyxb==1.2.6
xmltodict